Reorder conditions

+++ math/0x03-probability/poisson.py
@@ -15,7 +15,7 @@ class Poisson:
             if len(data) <= 2:
                 raise ValueError("data must contain multiple values")
             else:
-                self.lambtha = sum(data)/len(data)
+                self.lambtha = float(sum(data)/len(data))
         else:
             if lambtha < 0:
                 raise ValueError('lambtha must be a positive value')
